requirements.txt
python-telegram-bot==21.1
Flask==2.2.5
python-dotenv==1.0.0
requests==2.31.0